Modernised DFU Manifest
Instead of NSObject, so Objective-C classes, these are now structs. They're also Codable(s), so we don't need to do the JSON dance ourselves. I'm sure there are many more parts inside the library that deserve this same treatment, but we can't do it all at once. This is a small step.