Added Privacy Manifest for both Library and Example App
The APIs accessed are really under the purview of the ZIP Foundation library the library itself uses. But given that we might rely on older versions of library dependencies for the sake of convenience & compatibility (until something breaks), I thought it was prudent to add it ourselves as part of what our library does. After all, if we drop ZIPFoundation it's up to us, as library providers, to pick up the slack.